requirements.txt
python-dotenv
playwright
crewai
langchain
openai
